requirements.txt
streamlit
concrete-ml
scikit-learn
pandas
numpy
rpy2